+++ src/gateways/PaymentIntents.php
@@ -11,16 +11,18 @@ use Craft;
 use craft\commerce\base\Plan as BasePlan;
 use craft\commerce\base\RequestResponseInterface;
 use craft\commerce\base\SubscriptionResponseInterface;
+use craft\commerce\elements\Subscription;
 use craft\commerce\errors\SubscriptionException;
 use craft\commerce\models\payments\BasePaymentForm;
 use craft\commerce\models\PaymentSource;
-use craft\commerce\models\subscriptions\SubscriptionForm;
+use craft\commerce\models\subscriptions\SubscriptionForm as BaseSubscriptionForm;
 use craft\commerce\models\Transaction;
 use craft\commerce\Plugin as Commerce;
 use craft\commerce\stripe\base\SubscriptionGateway as BaseGateway;
 use craft\commerce\stripe\errors\PaymentSourceException;
 use craft\commerce\stripe\events\SubscriptionRequestEvent;
 use craft\commerce\stripe\models\forms\payment\PaymentIntent as PaymentForm;
+use craft\commerce\stripe\models\forms\Subscription as SubscriptionForm;
 use craft\commerce\stripe\models\PaymentIntent as PaymentIntentModel;
 use craft\commerce\stripe\Plugin as StripePlugin;
 use craft\commerce\stripe\responses\PaymentIntentResponse;
@@ -29,7 +31,6 @@ use craft\elements\User;
 use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
 use craft\web\View;
-use Stripe\Customer;
 use Stripe\PaymentIntent;
 use Stripe\PaymentMethod;
 use Stripe\Refund;
@@ -86,6 +87,7 @@ class PaymentIntents extends BaseGateway
         $defaults = [
             'gateway' => $this,
             'paymentForm' => $this->getPaymentFormModel(),
+            'scenario' => 'payment',
         ];
 
         $params = array_merge($defaults, $params);
@@ -201,6 +203,21 @@ class PaymentIntents extends BaseGateway
             throw new NotSupportedException('The currency “' . $transaction->paymentCurrency . '” is not supported!');
         }
 
+        // Are we dealing with a transaction that was created under the previous 'Charge' gateway?
+        if (substr($transaction->reference, 0, 3) === "ch_") {
+            try {
+                $request = [
+                    'charge' => $transaction->reference,
+                    'amount' => $transaction->paymentAmount * (10 ** $currency->minorUnit),
+                ];
+                $refund = Refund::create($request);
+
+                return $this->createPaymentResponseFromApiResource($refund);
+            } catch (\Exception $exception) {
+                return $this->createPaymentResponseFromError($exception);
+            }
+        }
+
         $stripePaymentIntent = PaymentIntent::retrieve($transaction->reference);
         $paymentIntentsService = StripePlugin::getInstance()->getPaymentIntents();
 
@@ -269,8 +286,9 @@ class PaymentIntents extends BaseGateway
      * @inheritdoc
      * @throws SubscriptionException if there was a problem subscribing to the plan
      */
-    public function subscribe(User $user, BasePlan $plan, SubscriptionForm $parameters): SubscriptionResponseInterface
+    public function subscribe(User $user, BasePlan $plan, BaseSubscriptionForm $parameters): SubscriptionResponseInterface
     {
+        /** @var SubscriptionForm $parameters */
         $customer = StripePlugin::getInstance()->getCustomers()->getCustomer($this->id, $user);
         $paymentMethods = PaymentMethod::all(['customer' => $customer->reference, 'type' => 'card']);
 
@@ -285,9 +303,13 @@ class PaymentIntents extends BaseGateway
 
         if ($parameters->trialDays !== null) {
             $subscriptionParameters['trial_period_days'] = (int)$parameters->trialDays;
+        } elseif ($parameters->trialEnd !== null) {
+            $subscriptionParameters['trial_end'] = (int)$parameters->trialEnd;;
         } else {
             $subscriptionParameters['trial_from_plan'] = true;
         }
+
+        $subscriptionParameters['expand'] = ['latest_invoice.payment_intent'];
 
         $event = new SubscriptionRequestEvent([
             'parameters' => $subscriptionParameters
@@ -322,8 +344,121 @@ class PaymentIntents extends BaseGateway
         return true;
     }
 
+    /**
+     * @inheritdoc
+     */
+    public function getBillingIssueDescription(Subscription $subscription): string
+    {
+        $subscriptionData = $this->_getExpandedSubscriptionData($subscription);
+        $intentData = $subscriptionData['latest_invoice']['payment_intent'];
+
+        if (in_array($subscriptionData['status'], ['incomplete', 'past_due', 'unpaid'])) {
+            switch ($intentData['status']) {
+                case 'requires_payment_method':
+                    return $subscription->hasStarted ? Craft::t('commerce-stripe', 'To resume the subscription, please provide a valid payment method.') : Craft::t('commerce-stripe', 'To start the subscription, please provide a valid payment method.');
+                case 'requires_action':
+                    return $subscription->hasStarted ? Craft::t('commerce-stripe', 'To resume the subscription, please complete 3DS authentication.') : Craft::t('commerce-stripe', 'To start the subscription, please complete 3DS authentication.');
+            }
+        }
+
+        return '';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getBillingIssueResolveFormHtml(Subscription $subscription): string
+    {
+        $subscriptionData = $this->_getExpandedSubscriptionData($subscription);
+        $intentData = $subscriptionData['latest_invoice']['payment_intent'];
+
+        if (in_array($subscriptionData['status'], ['incomplete', 'past_due', 'unpaid'])) {
+            $clientSecret = $intentData['client_secret'];
+            switch ($intentData['status']) {
+                case 'requires_payment_method':
+                case 'requires_confirmation':
+                    return $this->getPaymentFormHtml(['clientSecret' => $clientSecret]);
+                case 'requires_action':
+                    return $this->getPaymentFormHtml(['clientSecret' => $clientSecret, 'scenario' => '3ds']);
+            }
+        }
+
+        return '';
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getHasBillingIssues(Subscription $subscription): bool
+    {
+        $subscription = $this->refreshSubscriptionData($subscription);
+        $subscriptionData = $subscription->getSubscriptionData();
+        $intentData = $subscriptionData['latest_invoice']['payment_intent'];
+
+        return in_array($subscriptionData['status'], ['incomplete', 'past_due', 'unpaid']) && in_array($intentData['status'], ['requires_payment_method', 'requires_confirmation', 'requires_action']);
+    }
+
+
     // Protected methods
     // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    protected function handleWebhook(array $data)
+    {
+        switch ($data['type']) {
+            case 'invoice.payment_failed':
+                $this->handleInvoiceFailed($data);
+                break;
+        }
+
+        parent::handleWebhook($data);
+    }
+
+    /**
+     * Handle a failed invoice by updating the subscription data for the subscription it failed.
+     * @param $data
+     * @throws \Throwable
+     * @throws \craft\errors\ElementNotFoundException
+     * @throws \yii\base\Exception
+     */
+    protected function handleInvoiceFailed(array $data)
+    {
+        $stripeInvoice = $data['data']['object'];
+
+        // Sanity check
+        if ($stripeInvoice['paid']) {
+            return;
+        }
+
+        $subscriptionReference = $stripeInvoice['subscription'] ?? null;
+
+        if (!$subscriptionReference || !($subscription = Subscription::find()->anyStatus()->reference($subscriptionReference)->one())) {
+            Craft::warning('Subscription with the reference “' . $subscriptionReference . '” not found when processing webhook ' . $data['id'], 'stripe');
+
+            return;
+        }
+
+        $this->refreshSubscriptionData($subscription);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function handleSubscriptionUpdated(array $data)
+    {
+        // Fetch expanded data
+        $stripeSubscription = StripeSubscription::retrieve([
+            'id' => $data['data']['object']['id'],
+            'expand' => ['latest_invoice.payment_intent'],
+        ]);
+
+        // And nonchalantly replace it, before calling parent.
+        $data['data']['object'] = $stripeSubscription->jsonSerialize();
+
+        parent::handleSubscriptionUpdated($data);
+    }
 
     /**
      * @inheritdoc
@@ -391,6 +526,29 @@ class PaymentIntents extends BaseGateway
     }
 
     /**
+     * Refresh a subscription's data.
+     *
+     * @param $subscription
+     * @return Subscription
+     */
+    protected function refreshSubscriptionData(Subscription $subscription)
+    {
+        $stripeSubscription = StripeSubscription::retrieve([
+            'id' => $subscription->reference,
+            'expand' => ['latest_invoice.payment_intent']
+        ]);
+
+        $subscription->setSubscriptionData($stripeSubscription->jsonSerialize());
+        $this->setSubscriptionStatusData($subscription);
+        Craft::$app->getElements()->saveElement($subscription);
+
+        return $subscription;
+    }
+
+    // Private methods
+    // =========================================================================
+
+    /**
      * Confirm a payment intent and set the return URL.
      *
      * @param PaymentIntent $stripePaymentIntent
@@ -400,5 +558,27 @@ class PaymentIntents extends BaseGateway
         $stripePaymentIntent->confirm([
             'return_url' => UrlHelper::actionUrl('commerce/payments/complete-payment', ['commerceTransactionId' => $transaction->id, 'commerceTransactionHash' => $transaction->hash])
         ]);
+    }
+
+    /**
+     * Get the expanded subscription data, including payment intent for latest invoice.
+     *
+     * @param Subscription $subscription
+     * @return array
+     */
+    private function _getExpandedSubscriptionData(Subscription $subscription): array
+    {
+        $subscriptionData = $subscription->getSubscriptionData();
+
+        if (empty($subscriptionData['latest_invoice']['payment_intent'])) {
+            $stripeSubscription = StripeSubscription::retrieve([
+                'id' => $subscription->reference,
+                'expand' => ['latest_invoice.payment_intent']
+            ]);
+            $subscriptionData = $stripeSubscription->jsonSerialize();
+            $subscription->setSubscriptionData($subscriptionData);
+        }
+
+        return $subscriptionData;
     }
 }